requirements.txt
absl-py==0.8.1
astor==0.8.0
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
nltk==3.4.5
numpy==1.17.2
pandas==0.25.1
protobuf==3.10.0
python-dateutil==2.8.0
pytz==2019.3
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
tb-nightly==1.14.0a20190603
tensorflow-gpu==2.0.0b1
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
tqdm==4.36.1
Werkzeug==0.16.0
wrapt==1.11.2
xlrd==1.2.0
